third_party/blink/web_tests/external/wpt/css/selectors/parsing/parse-has.tentative.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Selectors: The relational pseudo-class (serialize :is() back unchanged)</title>
<link rel="author" title="Steinar H. Gunderson" href="mailto:sesse@chromium.org">
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/8356">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
  // This is split out from the non-tentative version of the test
  // because it depends on changes to serialization of :is() and :where()
  // that have not been standardized yet.
  test_valid_forgiving_selector(':has(:is(.a, 123))');
</script>
